app/requirements.txt
streamlit==1.38.0
pandas==2.2.2
numpy==1.26.4
plotly==5.24.1
python-pptx==0.6.23
reportlab==4.2.2
xlsxwriter==3.2.0
openpyxl==3.1.5
pillow==10.4.0
kaleido==0.2.1
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
